+++ app/javascript/react/components/calendar/CalendarGrid.tsx
@@ -1,10 +1,7 @@
-import { useCallback } from "react";
 import FullCalendar from "@fullcalendar/react";
 import dayGridPlugin from "@fullcalendar/daygrid";
 import interactionPlugin from "@fullcalendar/interaction";
-import type { EventSourceFuncArg, EventInput, EventClickArg, EventMountArg } from "@fullcalendar/core";
-import type { DateClickArg } from "@fullcalendar/interaction";
-import { fetchCalendarEvents } from "react/services/calendarApi";
+import { useCalendarHandlers } from "react/hooks/useCalendarHandlers";
 import { overlayStyle } from "./styles/overlayStyle";
 
 type CalendarGridProps = {
@@ -21,41 +18,12 @@ export default function CalendarGrid({
   calendarRef,
 }: CalendarGridProps) {
 
-  const handleEvents = useCallback(
-    async (
-      info: EventSourceFuncArg,
-      successCallBack: (events: EventInput[]) => void
-    ) => {
-      setLoading(true);
-      try {
-        const events = await fetchCalendarEvents(info.startStr, info.endStr);
-        successCallBack(events);
-      } finally {
-        setLoading(false);
-      }
-    },
-    [setLoading]
-  );
-
-  const handleDateClick = async (info: DateClickArg) => {
-    const date = info.dateStr;
-    const events = await fetchCalendarEvents(date, date);
-
-    if (events.length > 0) {
-      window.location.href = `/workouts/${events[0].id}`;
-    } else {
-      window.location.href = `/workouts/select_exercise?date=${date}`;
-    }
-  };
-
-  const handleEventClick = (info: EventClickArg) => {
-    window.location.href = `/workouts/${info.event.id}`;
-  };
-
-  const handleEventDidMount = (info: EventMountArg) => {
-    const frame = info.el.closest(".fc-daygrid-day-frame") as HTMLElement | null;
-    if (frame) frame.classList.add("fc-has-workout");
-  };
+  const {
+    handleEvents,
+    handleDateClick,
+    handleEventClick,
+    handleEventDidMount,
+  } = useCalendarHandlers(setLoading);
 
   return (
     <div style={{ marginTop: "1rem", position: "relative" }}>
